requirements.txt
# Web 框架
flask>=2.3.0
flask-cors>=4.0.0

# Web3
web3>=6.0.0

# HTTP 请求
requests>=2.28.0

# 数据处理
python-dotenv>=1.0.0

# MCP (如果有官方包)
# mcp>=0.1.0

# 可选：AI 分析
# openai>=1.0.0
